+++ server/db/models/index.ts
@@ -1,22 +1,12 @@
 import Sequelize from 'sequelize';
-import { DbInterface } from 'typings/DbInterface';
-import { ChurchFactory } from 'models/Church';
-import { UserFactory } from 'models/User';
-import { TaskFactory } from 'models/Task';
-import { RoleFactory } from 'models/Role';
-import configJson from '../config';
+import { DbInterface } from 'shared/SequelizeTypings/typings/DbInterface';
+import ChurchFactory from './Church';
+import UserFactory from './User';
+import TaskFactory from './Task';
+import RoleFactory from './Role';
 
-const env = process.env.NODE_ENV ? process.env.NODE_ENV : 'development';
-const config = configJson[env];
-config.logging = false;
-
-const createModels = (): DbInterface => {
-    const sequelize = new Sequelize(
-        config.database,
-        config.username,
-        config.password,
-        config,
-    );
+const createModels = (database, username, password, config): DbInterface => {
+    const sequelize = new Sequelize(database, username, password, config);
 
     const db: DbInterface = {
         sequelize,
@@ -32,6 +22,8 @@ const createModels = (): DbInterface => {
             db[modelName].associate(db);
         }
     });
+
+    // console.log(db.User);
 
     return db;
 };
