+++ server/routes/index.ts
@@ -1,12 +1,15 @@
-export {};
-const router = require('express').Router();
-module.exports = router;
+import express from 'express';
+
+const router = express.Router();
 
 router.use('/churches', require('./churches'));
 router.use('/users', require('./users'));
 
 router.use((req, res, next) => {
+    console.log('hello');
     const error = new Error('Not Found');
-    error.status = 404;
+    res.status(404);
     next(error);
 });
+
+module.exports = router;
