+++ shared/SequelizeTypings/models/RoleModel.ts
@@ -1,0 +1,14 @@
+import * as Sequelize from 'sequelize';
+
+export interface RoleAttributes {
+    id?: number;
+    name: string;
+    time: Date;
+    createdAt?: Date;
+    updatedAt?: Date;
+    churchId?: number;
+}
+
+export interface RoleInstance
+    extends Sequelize.Instance<RoleAttributes>,
+        RoleAttributes {}
