+++ shared/SequelizeTypings/models/index.d.ts
@@ -1,0 +1,4 @@
+export * from './ChurchModel';
+export * from './TaskModel';
+export * from './UserModel';
+export * from './RoleMOdel';
