+++ shared/SequelizeTypings/typings/SequelizeAttributes/index.d.ts
